services/api-manager/requirements.txt
# API Manager - Requirements
# Python 3.13+

# Web Framework
Quart==0.19.4
quart-cors==0.7.0

# Auth implemented separately for Quart

# Database - SQLAlchemy for schema, PyDAL for runtime
SQLAlchemy==2.0.36
alembic==1.14.0
pydal==20260110.1
psycopg2-binary==2.9.10
PyMySQL==1.1.0

# Authentication (keep for API tokens)
PyJWT==2.10.1
bcrypt==4.2.1

# Validation
email-validator==2.2.0

# Secrets Management
hvac==2.3.0
infisical-python==2.1.1
cryptography==43.0.3
boto3==1.35.78
google-cloud-secret-manager==2.20.0
azure-keyvault-secrets==4.8.0
azure-identity==1.17.1

# Cloud Provider SDKs
pylxd==2.3.2
google-cloud-compute==1.19.2
azure-mgmt-compute==37.1.0

# Async HTTP client (for MaaS OAuth)
aiohttp==3.11.11
httpx==0.27.0
oauthlib==3.2.2

# Template rendering
Jinja2==3.1.4
PyYAML==6.0.2

# ASGI Server
hypercorn==0.17.3

# Configuration
python-decouple==3.8

# Monitoring
prometheus-client==0.21.1

# Development
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.14.0
bandit==1.7.10
